+++ include/crpropa/Cosmology.h
@@ -17,64 +17,64 @@ namespace crpropa {
  @param hubbleParameter	dimensionless Hubble parameter, default = 0.673
  @param omegaMatter		matter parameter, default = 0.315
  */
-void setCosmologyParameters(double hubbleParameter, double omegaMatter);
+void setCosmologyParameters(long double hubbleParameter, long double omegaMatter);
 
 /**
  Hubble rate at given redshift
  H(z) = H0 * sqrt(omegaM * (1 + z)^3 + omegaL)
  */
-double hubbleRate(double redshift = 0);
+long double hubbleRate(long double redshift = 0);
 
 // Returns the dark energy density parameter
-double omegaL();
+long double omegaL();
 
 // Returns the matter density parameter
-double omegaM();
+long double omegaM();
 
 // Returns the hubble parameter
-double H0();
+long double H0();
 
 /**
  Redshift of a comoving object at a given comoving distance to an observer at z = 0.
  d_comoving(z) = c/H0 * int_0^z dz' / E(z')
  */
-double comovingDistance2Redshift(double distance);
+long double comovingDistance2Redshift(long double distance);
 
 /**
  Comoving distance between an observer at z = 0 and a comoving object at z.
  d_comoving(z) = c/H0 * int_0^z dz' / E(z')
  */
-double redshift2ComovingDistance(double redshift);
+long double redshift2ComovingDistance(long double redshift);
 
 /**
  Redshift of a comoving object at a given luminosity distance to an observer at z = 0.
  d_luminosity(z) = (1 + z) * d_comoving(z)
  */
-double luminosityDistance2Redshift(double distance);
+long double luminosityDistance2Redshift(long double distance);
 
 /**
  Luminosity distance between an observer at z = 0 and a comoving object at z.
  d_luminosity(z) = (1 + z) * d_comoving(z)
  */
-double redshift2LuminosityDistance(double redshift);
+long double redshift2LuminosityDistance(long double redshift);
 
 /**
  Redshift of a comoving object at a given light travel distance to an observer at z = 0.
  d_lighttravel(z) = c/H0 * int_0^z dz' / ((1 + z')  *  E(z'))
  */
-double lightTravelDistance2Redshift(double distance);
+long double lightTravelDistance2Redshift(long double distance);
 
 /**
  Light travel distance between an observer at z = 0 and a comoving object at z.
  d_lighttravel(z) = c/H0 * int_0^z dz' / ((1 + z')  *  E(z'))
  */
-double redshift2LightTravelDistance(double redshift);
+long double redshift2LightTravelDistance(long double redshift);
 
 // Conversion from comoving distance to light travel distance.
-double comoving2LightTravelDistance(double distance);
+long double comoving2LightTravelDistance(long double distance);
 
 // Conversion from light travel distance to comoving distance.
-double lightTravel2ComovingDistance(double distance);
+long double lightTravel2ComovingDistance(long double distance);
 
 /** @}*/
 } // namespace crpropa
